Vending drop and reload fully implemented

+++ src/Common/Vendor.Domain/DTO/Requests/LoadSpiralRequestDto.cs
@@ -5,5 +5,5 @@ public class LoadSpiralRequestDto
     public int Id { get; set; }
     public int ProductId { get; set; }
     public int Loads { get; set; }
-    public Double Price { get; set; }
+    public Decimal Price { get; set; }
 }
